🔨 Split `parser.ts`
